AyeSQL.Compiler (AyeSQL v1.1.4)

This module defines functions to compile AyeSQL language strings.

Types

docs()

@type docs() :: nil | binary()

Query documentation.

fragment()

@type fragment() :: binary()

Query fragment.

fragments()

@type fragments() :: [fragment() | param()]

Query fragments.

name()

@type name() :: nil | atom()

Query name.

param()

@type param() :: atom()

Query parameter.

queries()

@type queries() :: [query()]

Queries.

query()

@type query() :: {name(), docs(), fragments()}

Query.

Functions

compile_queries(contents, options \\ [])

@spec compile_queries(binary(), AyeSQL.Lexer.options()) :: [Macro.t()] | no_return()

Compiles the contents of a file or string into valid AyeSQL queries.

compile_query(contents, options \\ [])

@spec compile_query(binary(), AyeSQL.Lexer.options()) :: Macro.t() | no_return()

Compiles a single query from the contents of a string.

eval_query(contents, options \\ [])

@spec eval_query(binary(), AyeSQL.Lexer.options()) ::
  (AyeSQL.Core.parameters(), AyeSQL.Core.options() ->
     {:ok, AyeSQL.Query.t() | term()} | {:error, AyeSQL.Error.t() | term()})
  | no_return()

Evaluates the contents of a string to an anonymous function with a query that receives parameters and options.